requirements.txt
Flask==3.0.2
selenium==4.18.1
undetected-chromedriver==3.5.5
pandas==2.2.1
openpyxl==3.1.2
python-dotenv==1.0.1
SQLAlchemy==2.0.28
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
requests==2.31.0
PySocks==1.7.1
python-socketio==5.11.1
Flask-SocketIO==5.3.6 